Refactor: Decouple caching logic from episode metadata
This commit refactors the media caching data structures and logic to decouple them from domain-specific metadata like episode titles. The `EpisodeMetadata` class now serves strictly as a unique key (`aid`, `bvid`, `cid`) for a cached item, and the `MediaCacheSave` class has been updated to use this `key`.

This change simplifies the caching system by removing the responsibility of storing descriptive information (like titles) from the core cache entity. The `MediaCacheDataSource` API is now more generic, using `get`, `save`, and `delete` operations based on the `EpisodeMetadata` key.

Specific changes:
- In `core/model`:
    - Renamed `MediaCacheSave.origin` to `MediaCacheSave.key`.
    - Changed `EpisodeMetadata.aid` from `Long` to `String` to align with data source types.
    - Removed the `title` field from `EpisodeMetadata`, making it a pure identifier.
    - Replaced `MediaCachePartMetadata.downloadId` with `fullPath` to store the complete file path directly.

- In `core/datasource/local/MediaCacheDataSource`:
    - Replaced `listFlow` with `allCachesFlow`.
    - Refactored methods to be key-based (`get`, `save`, `delete`) instead of querying by individual fields or updating parts of the data.

- In `core/domain` UseCases (`GetPgcEpisodeInfoUseCase`, `GetPugvEpisodeInfoUseCase`, `GetUgcEpisodeInfoUseCase`, `GetCachedEpisodeStateUseCase`):
    - Updated logic to use `allCachesFlow` and match cached items using the new `key` property.

- In feature modules (`cache`, `search`, `video-player`):
    - Adapted ViewModels and Composables to the new data source API and models.
    - Updated navigation arguments and cache creation logic to use `String` for `aid`.

+++ app/src/commonMain/kotlin/com/imcys/bilibilias/di/NavigationModule.kt
@@ -51,7 +51,7 @@ val NavigationModule = module {
             entry<CacheRoute> {
                 CacheScreen(
                     navigationToPlayer = { aid, bvid, cid ->
-                        backStack.navigate(PlayerRoute(aid, bvid, cid))
+                        backStack.navigate(PlayerRoute(aid.toLong(), bvid, cid))
                     },
                 )
             }
